Testing branch - testing branch to fix bugs

+++ server/src/schemas/resolvers.ts
@@ -1,12 +1,4 @@
-import {
-  User,
-  Profile,
-  Prompt,
-  Story,
-  Comment,
-  Vote,
-} from "../models/index.js";
-
+import { User, Profile, Prompt, Story, Comment, Vote } from "../models/index.js";
 import { signToken } from "../utils/auth.js";
 
 const resolvers = {
@@ -32,7 +24,6 @@ const resolvers = {
       const profile = await Profile.findOne({ user: context.user._id })
         .populate("user")
         .populate("followers")
-        .populate("following")
         .populate("sharedStories")
         .populate("branchedStories")
         .populate("likedStories");
@@ -54,9 +45,9 @@ const resolvers = {
     // Login an existing user and return a JWT token
     login: async (
       _parent: any,
-      { email, password }: { email: string; password: string }
+      { username, password }: { username: string; password: string }
     ) => {
-      const user = await User.findOne({ email });
+      const user = await User.findOne({ username });
       if (!user) {
         throw new Error("Can't find this user");
       }
@@ -74,12 +65,13 @@ const resolvers = {
     addUser: async (
       _: any,
       {
+        fullName,
         username,
         email,
         password,
-      }: { username: string; email: string; password: string }
+      }: { fullName: string; username: string; email: string; password: string }
     ) => {
-      const user = await User.create({ username, email, password });
+      const user = await User.create({ fullName, username, email, password });
       await Profile.create({ user: user._id });
       const token = signToken(user.username, user.email, user._id);
       return { token, user };
@@ -127,9 +119,7 @@ const resolvers = {
         parentStory: originalStory._id,
       });
 
-      originalStory.branches.push(
-        branchedStory._id as (typeof originalStory.branches)[0]
-      );
+      originalStory.branches.push(branchedStory._id as typeof originalStory.branches[0]);
       await originalStory.save();
 
       await Profile.findOneAndUpdate(
@@ -183,22 +173,16 @@ const resolvers = {
       return comment;
     },
 
-    // Delete a story and related comments (but not branched stories)
-    deleteStory: async (
+     // Delete a story and related comments (but not branched stories)
+     deleteStory: async (
       _: any,
       { storyId }: { storyId: string },
       context: any
     ) => {
       if (!context.user) throw new Error("You need to be logged in!");
 
-      const story = await Story.findOne({
-        _id: storyId,
-        author: context.user._id,
-      });
-      if (!story)
-        throw new Error(
-          "Story not found or you're not authorized to delete it"
-        );
+      const story = await Story.findOne({ _id: storyId, author: context.user._id });
+      if (!story) throw new Error("Story not found or you're not authorized to delete it");
 
       // Delete associated comments
       await Comment.deleteMany({ story: storyId });
@@ -229,10 +213,7 @@ const resolvers = {
     // Vote for a story (upvote/downvote)
     voteStory: async (
       _: any,
-      {
-        storyId,
-        voteType,
-      }: { storyId: string; voteType: "upvote" | "downvote" },
+      { storyId, voteType }: { storyId: string; voteType: "upvote" | "downvote" },
       context: any
     ) => {
       if (!context.user) throw new Error("You need to be logged in!");
@@ -255,42 +236,6 @@ const resolvers = {
       });
 
       return newVote;
-    },
-
-    // Follow another user
-    followUser: async (
-      _: any,
-      { profileId }: { profileId: string },
-      context: any
-    ) => {
-      if (!context.user) throw new Error("You need to be logged in!");
-
-      const updatedProfile = await Profile.findByIdAndUpdate(
-        profileId,
-        { $addToSet: { followers: context.user._id } },
-        { new: true }
-      );
-
-      if (!updatedProfile) throw new Error("Profile not found");
-      return updatedProfile;
-    },
-
-    // Unfollow another user
-    unfollowUser: async (
-      _: any,
-      { profileId }: { profileId: string },
-      context: any
-    ) => {
-      if (!context.user) throw new Error("You need to be logged in!");
-
-      const updatedProfile = await Profile.findByIdAndUpdate(
-        profileId,
-        { $pull: { followers: context.user._id } },
-        { new: true }
-      );
-
-      if (!updatedProfile) throw new Error("Profile not found");
-      return updatedProfile;
     },
   },
 };
